Implementation using COPY is working
