+++ Cargo.toml
@@ -9,15 +9,12 @@ chrono = "0.4.11"
 dialoguer = "0.6.2"
 indicatif = "0.14.0"
 log = "0.4.8"
-postgres = "0.17.2"
+postgres = "0.17.5"
 resolve = "0.2.0"
 rust_decimal = { version = "1.4.1", features = ["postgres"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_yaml = "0.8"
 simple-logging = "2.0.2"
 
-# Include all default features and also "postgres" implementation of Decimal
-#[dependencies.rust_decimal]
-#version = "1.4.1"
-#default-features = true
-#features = ["postgres"]
+[profile.release]
+lto = true
